+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Added humanized form of `#inspect` and `#to_s` for better readability when outputting model instances
 - Added internal logger
+- Added `#update` method to all models to update attributes post-creation
 
 ### Changed
 - **Breaking change**: Renamed the cache to `:all` and the iVar` to `@all` for clarity